Refactor : Moved code to models
